requirements.txt
Python 3.8 or higher

Pillow (optional, for saving images)
